# gah-api/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error"
      }
    },
    "schemas": {
      "AgentConfig": {
        "properties": {
          "api_base_url": {
            "nullable": true,
            "type": "string"
          },
          "api_key": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "system_prompt": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "api_key",
          "model",
          "provider"
        ],
        "type": "object"
      },
      "ChatMessage": {
        "properties": {
          "content": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tool_call_id": {
            "nullable": true,
            "type": "string"
          },
          "tool_calls": {
            "nullable": true
          }
        },
        "required": [
          "content",
          "role"
        ],
        "type": "object"
      },
      "CreateSessionRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/AgentConfig"
          }
        },
        "required": [
          "config"
        ],
        "type": "object"
      },
      "CreateTokenRequest": {
        "properties": {
          "label": {
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "type": "object"
      },
      "CreateTokenResponse": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "label",
          "token"
        ],
        "type": "object"
      },
      "Error": {
        "description": "Error information from a response.",
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "request_id"
        ],
        "type": "object"
      },
      "ProviderKind": {
        "enum": [
          "openai",
          "anthropic",
          "openrouter",
          "ollama",
          "custom"
        ],
        "type": "string"
      },
      "SendPromptRequest": {
        "properties": {
          "prompt": {
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "type": "object"
      },
      "SendPromptResponse": {
        "properties": {
          "message_count": {
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "output": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageInfo"
          }
        },
        "required": [
          "message_count",
          "output",
          "usage"
        ],
        "type": "object"
      },
      "SessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "sessions"
        ],
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/AgentConfig"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "message_count": {
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "config",
          "created_at",
          "id",
          "message_count",
          "messages",
          "updated_at"
        ],
        "type": "object"
      },
      "UsageInfo": {
        "properties": {
          "input_tokens": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "output_tokens": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "input_tokens",
          "output_tokens"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "gah-api",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/v1/sessions": {
      "get": {
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            },
            "description": "successful operation"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "successful creation"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/sessions/{id}": {
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "resource updated"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "successful operation"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/sessions/{id}/prompt": {
      "post": {
        "operationId": "send_prompt",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendPromptResponse"
                }
              }
            },
            "description": "successful operation"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/sessions/{id}/stream": {
      "get": {
        "description": "receive a stream of JSON-encoded [`AgentEvent`] frames, ending with a `done` (or `error`) event.",
        "operationId": "session_stream",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Negotiating protocol upgrade from HTTP/1.1 to WebSocket"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Stream a prompt over a websocket. Send one text frame with the prompt;",
        "x-dropshot-websocket": {}
      }
    },
    "/v1/tokens": {
      "post": {
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponse"
                }
              }
            },
            "description": "successful creation"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  }
}